// src/app/core/models/point-detail/point-detail.model.ts
export interface MaintanenceResponseModel {
    alertLevel: number;
    id: string;
    mainSensorId: string;
    mode: string;
    name: string;
    number: number;
    recordingCycle: string;
    recordingCycleInitial:string;
    transmissionCycle: string;
    transmissionCycleInitial: string;
    updatedDate: string;
}

export interface MaintanenceRequestModel {
    alertLevel?: number;
    id?: string;
    mainSensorId?: string;
    mode?: string;
    name?: string;
    number?: number;
    recordingCycle?: string;
    recordingCycleInitial?:string;
    transmissionCycle?: string;
    transmissionCycleInitial?: string;
    updatedDate?: string;
}